chore(swift): update swift-sdk seed
