+++ seed/swift-sdk/exhaustive/Snippets/Example23.swift
@@ -7,9 +7,10 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.object.getAndReturnWithDatetimeLikeString(request: ObjectWithDatetimeLikeString(
-        datetimeLikeString: "datetimeLikeString",
-        actualDatetime: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)
+    _ = try await client.endpoints.object.getAndReturnWithUnknownField(request: ObjectWithUnknownField(
+        unknown: .object([
+            "key": .string("value")
+        ])
     ))
 }
