+++ seed/swift-sdk/exhaustive/Snippets/Example28.swift
@@ -7,10 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.params.getWithQuery(
-        query: "query",
-        number: 1
-    )
+    _ = try await client.endpoints.params.getWithPath(param: "param")
 }
 
 try await main()
