+++ seed/swift-sdk/exhaustive/Snippets/Example30.swift
@@ -7,9 +7,9 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.params.getWithPathAndQuery(
-        param: "param",
-        query: "query"
+    _ = try await client.endpoints.params.getWithQuery(
+        query: "query",
+        number: 1
     )
 }
