+++ seed/swift-sdk/exhaustive/Snippets/Example33.swift
@@ -7,9 +7,9 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.params.uploadWithPath(
-        param: "upload-path",
-        request: Data("data".utf8)
+    _ = try await client.endpoints.params.modifyWithPath(
+        param: "param",
+        request: "string"
     )
 }
