+++ seed/swift-sdk/exhaustive/Snippets/Example34.swift
@@ -7,7 +7,10 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.primitive.getAndReturnString(request: "string")
+    _ = try await client.endpoints.params.modifyWithPath(
+        param: "param",
+        request: "string"
+    )
 }
 
 try await main()
