+++ seed/swift-sdk/exhaustive/Snippets/Example35.swift
@@ -7,7 +7,10 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.primitive.getAndReturnInt(request: 1)
+    _ = try await client.endpoints.params.uploadWithPath(
+        param: "upload-path",
+        request: Data("data".utf8)
+    )
 }
 
 try await main()
