+++ seed/swift-sdk/exhaustive/Snippets/Example36.swift
@@ -7,7 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.primitive.getAndReturnLong(request: 1000000)
+    _ = try await client.endpoints.primitive.getAndReturnString(request: "string")
 }
 
 try await main()
