+++ seed/swift-sdk/exhaustive/Snippets/Example40.swift
@@ -7,7 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.primitive.getAndReturnDate(request: CalendarDate("2023-01-15")!)
+    _ = try await client.endpoints.primitive.getAndReturnBool(request: true)
 }
 
 try await main()
