+++ seed/swift-sdk/exhaustive/Snippets/Example44.swift
@@ -7,12 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.union.getAndReturnUnion(request: Animal.dog(
-        .init(
-            name: "name",
-            likesToWoof: true
-        )
-    ))
+    _ = try await client.endpoints.primitive.getAndReturnBase64(request: "SGVsbG8gd29ybGQh")
 }
 
 try await main()
