+++ seed/swift-sdk/exhaustive/Snippets/Example46.swift
@@ -7,7 +7,12 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.urls.noEndingSlash()
+    _ = try await client.endpoints.union.getAndReturnUnion(request: Animal.dog(
+        .init(
+            name: "name",
+            likesToWoof: true
+        )
+    ))
 }
 
 try await main()
