+++ seed/swift-sdk/exhaustive/Snippets/Example47.swift
@@ -7,7 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.urls.withEndingSlash()
+    _ = try await client.endpoints.urls.withMixedCase()
 }
 
 try await main()
