+++ seed/swift-sdk/exhaustive/Snippets/Example48.swift
@@ -7,7 +7,7 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.endpoints.urls.withUnderscores()
+    _ = try await client.endpoints.urls.noEndingSlash()
 }
 
 try await main()
