+++ seed/swift-sdk/exhaustive/Snippets/Example53.swift
@@ -7,7 +7,9 @@ private func main() async throws {
         token: "<token>"
     )
 
-    _ = try await client.noReqBody.getWithNoRequestBody()
+    _ = try await client.noAuth.postWithNoAuth(request: .object([
+        "key": .string("value")
+    ]))
 }
 
 try await main()
