+++ seed/swift-sdk/exhaustive/Sources/Resources/Endpoints/Object/ObjectClient.swift
@@ -67,6 +67,16 @@ public final class ObjectClient: Sendable {
         )
     }
 
+    public func getAndReturnWithUnknownField(request: ObjectWithUnknownField, requestOptions: RequestOptions? = nil) async throws -> ObjectWithUnknownField {
+        return try await httpClient.performRequest(
+            method: .post,
+            path: "/object/get-and-return-with-unknown-field",
+            body: request,
+            requestOptions: requestOptions,
+            responseType: ObjectWithUnknownField.self
+        )
+    }
+
     /// Tests that string fields containing datetime-like values are NOT reformatted.
     /// The datetimeLikeString field should preserve its exact value "2023-08-31T14:15:22Z"
     /// without being converted to "2023-08-31T14:15:22.000Z".
